requirements.txt
RateMyProfessorAPI 
textblob